abstract out performValidations
